refactor digestio into lib
