+++ config/initializers/digest_io.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require_dependency Rails.root / 'lib' / 'digest_io'
